+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- Countries: show states for Mexico
+
 ### Changed
 
 - upgraded Flutter to stable v3.27.1